+++ resolwe_bio/tests/__init__.py
@@ -1,9 +1,0 @@
-"""
-**************
-Procesor Tests
-**************
-
- .. automodule:: server.tests.processor.base
- .. automodule:: server.tests.processor.utils
-
-"""